requirements.txt
openai
httpx==0.27.2
python-dotenv
tabulate
fastapi
uvicorn[standard]
python-multipart
aiofiles
